+++ textwire_test.go
@@ -65,8 +65,9 @@ func TestErrorHandling(t *testing.T) {
 	}{
 		{`{{ 1 + "a" }}`, fail.New(1, "", "evaluator", fail.ErrTypeMismatch, object.INT_OBJ, "+", object.STR_OBJ), nil},
 		{`@use("sometemplate")`, fail.New(1, "", "evaluator", fail.ErrUseStmtMustHaveProgram), nil},
-		{`{{ loop = "test" }}`, fail.New(1, "", "evaluator", fail.ErrLoopVariableIsReserved), nil},
-		{`{{ loop }}`, fail.New(0, "", "evaluator", fail.ErrLoopVariableIsReserved), map[string]any{"loop": "test"}},
+		{`{{ loop = "test" }}`, fail.New(1, "", "evaluator", fail.ErrReservedVariables), nil},
+		{`{{ global = "test" }}`, fail.New(1, "", "evaluator", fail.ErrReservedVariables), nil},
+		{`{{ loop }}`, fail.New(0, "", "evaluator", fail.ErrReservedVariables), map[string]any{"loop": "test"}},
 		{`{{ n = 1; n = "test" }}`, fail.New(1, "", "evaluator", fail.ErrVariableTypeMismatch, "n", object.INT_OBJ, object.STR_OBJ), nil},
 		{`{{ obj = {}; obj.name }}`, fail.New(1, "", "evaluator", fail.ErrPropertyNotFound, "name", object.OBJ_OBJ), nil},
 		{`{{ {}.test }}`, fail.New(1, "", "evaluator", fail.ErrPropertyNotFound, "test", object.OBJ_OBJ), nil},