+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v2.8.0 (2026-01-25)
 - 🐛 Fixed incorrect file path in error messages when error happens inside of `@insert` directive.
-- ⚠️ Variable `global` is now reserved. Potential breaking change.
+- ⚠️ BREAKING CHANGE! variable `global` is now reserved.
 
 ## v2.7.1 (2026-01-23)
 - 🐛 Fixed incorrect file path in error messages when error happens inside of components