+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v2.8.0 (2026-01-25)
 - 🐛 Fixed incorrect file path in error messages when error happens inside of `@insert` directive.
 - ⚠️ BREAKING CHANGE! variable `global` is now reserved.
+- ✨ Improve memory and performance.
 
 ## v2.7.1 (2026-01-23)
 - 🐛 Fixed incorrect file path in error messages when error happens inside of components